+++ CHANGELOG.md
@@ -131,6 +131,8 @@ BUG FIXES:
     the newest versions of Docker. [GH-2874]
   - provisioners/puppet: Append default module path to the module paths
     always. [GH-2677]
+  - provisioners/salt: Setting pillar data doesn't require `deep_merge`
+    plugin anymore. [GH-2348]
   - provisioners/salt: Options can now set install type and install args.
     [GH-2766]
   - provisioners/salt: Fix case when salt would say "options only allowed